Move find executable
