+++ perun/utils/__init__.py
@@ -122,26 +122,3 @@ def get_project_elf_executables(root: str = ".", only_not_stripped: bool = False
         binaries += list(get_directory_elf_executables(build_dir, only_not_stripped))
 
     return binaries
-
-
-def find_executable(cmd: Optional[str]) -> Optional[str]:
-    """Check if the supplied cmd is executable and find its real path
-    (i.e. absolute path with resolved symlinks)
-
-    :param str cmd: the command to check
-
-    :return str: resolved command path
-    """
-    # Ignore invalid paths
-    if cmd is None:
-        return None
-
-    # shutil.which checks:
-    # 1) files with relative / absolute paths specified
-    # 2) files accessible through the user PATH environment variable
-    # 3) that the file is indeed accessible and executable
-    cmd = shutil.which(cmd)
-    if cmd is None:
-        return None
-    # However, we still want to resolve the real path of the file
-    return os.path.realpath(cmd)
